+++ ReleaseNotes.md
@@ -2,6 +2,8 @@
 
 ## vNext
 
+- Added new version for static encryptor to encrypt/decrypt strings.
+
 ## 0.5.0
 
 - Fixed the build script to include symbols.